+++ .gitignore
@@ -7,6 +7,7 @@
 */.idea/*
 **/.vscode/*
 **/out/
+&&/node_modules/
 **/logs/
 corda-nodeinfo/corda-nodeinfo.iml
 *~*
