--- Tokens/tokentofriend/clients/src/main/kotlin/com/tokentofriend/webserver/Server.kt
@@ -1,22 +1,0 @@
-package com.tokentofriend.webserver
-
-import org.springframework.boot.Banner
-import org.springframework.boot.SpringApplication
-import org.springframework.boot.WebApplicationType.SERVLET
-import org.springframework.boot.autoconfigure.SpringBootApplication
-
-/**
- * Our Spring Boot application.
- */
-@SpringBootApplication
-private open class Starter
-
-/**
- * Starts our Spring Boot application.
- */
-fun main(args: Array<String>) {
-    val app = SpringApplication(Starter::class.java)
-    app.setBannerMode(Banner.Mode.OFF)
-    app.webApplicationType = SERVLET
-    app.run(*args)
-}
