+++ Tokens/tokentofriend/clients/src/main/kotlin/net/corda/samples/tokentofriend/webserver/Server.kt
@@ -1,0 +1,22 @@
+package net.corda.samples.tokentofriend.webserver
+
+import org.springframework.boot.Banner
+import org.springframework.boot.SpringApplication
+import org.springframework.boot.WebApplicationType.SERVLET
+import org.springframework.boot.autoconfigure.SpringBootApplication
+
+/**
+ * Our Spring Boot application.
+ */
+@SpringBootApplication
+private open class Starter
+
+/**
+ * Starts our Spring Boot application.
+ */
+fun main(args: Array<String>) {
+    val app = SpringApplication(Starter::class.java)
+    app.setBannerMode(Banner.Mode.OFF)
+    app.webApplicationType = SERVLET
+    app.run(*args)
+}
