--- Tokens/tokentofriend/contracts/src/test/kotlin/com/tokentofriend/contracts/StateTests.kt
@@ -1,18 +1,0 @@
-package com.tokentofriend.contracts
-
-import com.tokentofriend.states.CustomTokenState
-import groovy.util.GroovyTestCase.assertEquals
-import net.corda.testing.node.MockServices
-import org.junit.Test
-
-class StateTests {
-    private val ledgerServices = MockServices()
-
-    @Test
-    fun hasMessageFieldOfCorrectType() {
-        // Does the message field exist?
-        CustomTokenState::class.java.getDeclaredField("message")
-        // Is the message field of the correct type?
-        assertEquals(CustomTokenState::class.java.getDeclaredField("message").type, String::class.java)
-    }
-}
