+++ Tokens/tokentofriend/gradle.properties
@@ -1,4 +1,4 @@
-name=Test
+name=Tokentofriend Cordapp
 group=com.tokentofriend
-version=0.1
+version=1.0
 kotlin.incremental=false
