+++ Tokens/tokentofriend/runServers.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+./gradlew runOperatoreServer &
+./gradlew runUSWest1Server
+
